+++ Changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on Keep a Changelog,
 and this project adheres to Semantic Versioning.
 
-[1-4] - 2025-10-18
+[1-13] - 2025-10-18
 
 Added
 
@@ -17,4 +17,4 @@ Added
 - Enable or disable it with a toggle.
 - Open the file in the default editor.
 - Remove the file from the management list.
-- Centralizing Configuration
+- Centralizing Configuration & logs